did a major code refactor. using react context for authentication now.
