+++ src/components/auth/hooks/useAuthFlow.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { AuthContext } from "@/components/auth/contexts/AuthContext.tsx";
+
+export function useAuthFlow() {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error("useAuth must be used within an AuthProvider.");
+  }
+  return context;
+}
